Spring-Boot 1.0

+++ src/main/java/web/service/UserService.java
@@ -7,13 +7,13 @@ import java.util.List;
 public interface UserService {
     public List<User> getAllUsers();
 
-    public User show(Integer id);
+    public User show(Long id);
 
-    public void update(Integer id, User updateUser);
+    public void update(Long id, User updateUser);
 
     public void save(User user);
 
-    public void delete(Integer id);
+    public void delete(Long id);
 
     public User getUserByName(String name);
 }
